// app/models/firebaseDailyPrayerTimes.ts
// src/models/firebaseDailyPrayerTimes.ts
export interface FirebaseDailyPrayerTimes {
    fajr_adhan: string;
    fajr_iqama: string;
    dhuhr_adhan: string;
    dhuhr_iqama: string;
    asr_adhan: string;
    asr_iqama: string;
    maghrib_adhan: string;
    maghrib_iqama: string;
    isha_adhan: string;
    isha_iqama: string;
    jumaa_khutba: string;
    jumaa_salah: string;
  }
  